fix argument parsing for 64 bits
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@25549 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/examples/sendrecv/sr_bench.c
@@ -96,13 +96,13 @@ int main(int argc, char	**argv)
   
   for(i=1 ; i<argc ; i+=2) {
     if (!strcmp(argv[i], "-S")) {
-      start_len = atoi(argv[i+1]);
+      start_len = atoll(argv[i+1]);
     }
     else if (!strcmp(argv[i], "-E")) {
-      end_len = atoi(argv[i+1]);
+      end_len = atoll(argv[i+1]);
     }
     else if (!strcmp(argv[i], "-I")) {
-      increment = atoi(argv[i+1]);
+      increment = atoll(argv[i+1]);
     }
     else if (!strcmp(argv[i], "-M")) {
       multiplier = atof(argv[i+1]);
